--- src/app/(auth)/log-in/_components/ConsumerLogInForm.tsx
@@ -1,128 +1,0 @@
-'use client';
-
-import Button from '@/components/Buttons/Button';
-import React, { useState } from 'react';
-import { Controller, FormProvider, useForm } from 'react-hook-form';
-import { IoMdEye, IoMdEyeOff } from 'react-icons/io';
-import Input from '@/components/Inputs/Input';
-import AppleLogo from '../assets/AppleLogo';
-import GoogleLogo from '../assets/GoogleLogo';
-import Checkbox from '@/components/Inputs/Checkbox';
-import { useConsumerLogin } from '@/querys/ConsumerQuerys';
-import Link from 'next/link';
-
-const ConsumerLogInForm = () => {
-    const methods = useForm({
-        defaultValues: {
-            email: '',
-            password: '',
-            rememberMe: 'off',
-            saveID: 'off'
-        }
-    });
-    const [showPassword, setShowPassword] = useState(false);
-    const { control } = methods;
-    const { mutate: login } = useConsumerLogin();
-
-    const onSubmit = (data: any) => {
-        const { email, password } = data;
-
-        login({ email, password });
-    };
-
-    const handleClickIcon = () => {
-        setShowPassword(!showPassword);
-    };
-    return (
-        <FormProvider {...methods}>
-            <form onSubmit={methods.handleSubmit(onSubmit)}>
-                <div className="flex flex-col w-96">
-                    <div className="flex flex-col gap-6 h-56 mb-20">
-                        <Controller
-                            name="email"
-                            control={control}
-                            rules={{
-                                required: 'Email is required',
-                                pattern: {
-                                    value: /^[^\s@]+@[^\s@]+\.[^\s@]+$/,
-                                    message: 'Please enter a valid email address.'
-                                },
-                                maxLength: {
-                                    value: 50,
-                                    message: 'Email must be 50 characters or less.'
-                                }
-                            }}
-                            render={({ field, fieldState }) => (
-                                <Input
-                                    {...field}
-                                    label="ID"
-                                    placeholder="ex)abcd@gmail.com"
-                                    state={fieldState.error ? 'error' : 'default'}
-                                    validationMessage={fieldState.error?.message}
-                                />
-                            )}
-                        />
-                        <Controller
-                            name="password"
-                            control={control}
-                            rules={{
-                                required: 'Password is required',
-
-                                maxLength: {
-                                    value: 30,
-                                    message: 'Password must be 30 characters or less.'
-                                }
-                            }}
-                            render={({ field, fieldState }) => (
-                                <Input
-                                    {...field}
-                                    label="Password"
-                                    type={showPassword ? 'text' : 'password'}
-                                    placeholder="ex)123@abcd"
-                                    icon={showPassword ? <IoMdEye /> : <IoMdEyeOff />}
-                                    handleClickIcon={handleClickIcon}
-                                    state={fieldState.error ? 'error' : 'default'}
-                                    validationMessage={fieldState.error?.message}
-                                />
-                            )}
-                        />
-                        <div className="flex gap-2">
-                            <Checkbox name="rememberMe">remember-me</Checkbox>
-                            <Checkbox name="saveID">save ID</Checkbox>
-                        </div>
-                    </div>
-
-                    <div className="flex flex-col w-96 h-24 gap-4 mb-16">
-                        <Button priority="primary" size="md" fullWidth>
-                            Login
-                        </Button>
-                        <div className="flex w-full justify-between text-gray-500 text-[16px]">
-                            <p className="underline">Forgot password?</p>
-                            <p>
-                                Don't have an account?
-                                <Link href={`/sign-up/customer`}>
-                                    <span className="text-sub500 underline">Join</span>
-                                </Link>
-                            </p>
-                        </div>
-                    </div>
-                    <div className="flex flex-col w-96 h-28 gap-6">
-                        <div className="relative flex items-center">
-                            <hr className="w-full border-gray-300" />
-                            <span className="absolute bg-white px-3 left-1/2 top-1/2 transform -translate-x-1/2 -translate-y-1/2 text-gray-500 text-[16px] whitespace-nowrap">
-                                Login/Join with SNS account
-                            </span>
-                        </div>
-
-                        <div className="flex gap-10 justify-center">
-                            <AppleLogo />
-                            <GoogleLogo />
-                        </div>
-                    </div>
-                </div>
-            </form>
-        </FormProvider>
-    );
-};
-
-export default ConsumerLogInForm;
