+++ src/app/(auth)/log-in/agent/page.tsx
@@ -1,0 +1,14 @@
+import LogInLayout from '../_components/LogInLayout';
+import AgentLogInForm from './_components/AgentLogInForm';
+
+const AgentLogInPage = () => {
+    return (
+        <main className="flex flex-col items-center justify-center h-screen overflow-y-scroll">
+            <LogInLayout>
+                <AgentLogInForm />
+            </LogInLayout>
+        </main>
+    );
+};
+
+export default AgentLogInPage;
