--- src/app/(auth)/log-in/assets/AppleLogo.tsx
@@ -1,22 +1,0 @@
-import React from 'react';
-
-const AppleLogo = () => {
-    return (
-        <svg width="56" height="56" viewBox="0 0 56 56" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <rect width="56" height="56" rx="28" fill="black" />
-            <g clipPath="url(#clip0_1593_15753)">
-                <path
-                    d="M39.2482 23.3136C37.2819 24.5286 36.0675 26.6116 36.0675 28.9259C36.0675 31.5296 37.6289 33.9018 40 34.8854C39.5373 36.3898 38.8434 37.7784 37.9759 39.0513C36.7036 40.8449 35.3735 42.6964 33.4072 42.6964C31.441 42.6964 30.8627 41.5392 28.5494 41.5392C26.294 41.5392 25.4843 42.7543 23.6337 42.7543C21.7831 42.7543 20.5108 41.0764 19.0651 38.9934C17.1566 36.1005 16.0578 32.7446 16 29.2152C16 23.4872 19.7012 20.4206 23.4024 20.4206C25.3687 20.4206 26.988 21.6935 28.2024 21.6935C29.359 21.6935 31.2096 20.3628 33.4072 20.3628C35.7205 20.3049 37.9181 21.4042 39.2482 23.3136ZM32.3663 17.9327C33.3494 16.7755 33.8699 15.329 33.9277 13.8247C33.9277 13.6511 33.9277 13.4197 33.8699 13.2461C32.1928 13.4197 30.6313 14.2297 29.5325 15.5026C28.5494 16.6019 27.9711 17.9905 27.9133 19.4949C27.9133 19.6685 27.9133 19.842 27.9711 20.0156C28.0867 20.0156 28.2602 20.0735 28.3759 20.0735C29.9373 19.9578 31.3831 19.1477 32.3663 17.9327Z"
-                    fill="white"
-                />
-            </g>
-            <defs>
-                <clipPath id="clip0_1593_15753">
-                    <rect width="24" height="29.5082" fill="white" transform="translate(16 13.2461)" />
-                </clipPath>
-            </defs>
-        </svg>
-    );
-};
-
-export default AppleLogo;
