--- src/app/(auth)/log-in/assets/GoogleLogo.tsx
@@ -1,28 +1,0 @@
-import React from 'react';
-
-const GoogleLogo = () => {
-    return (
-        <svg width="56" height="56" viewBox="0 0 56 56" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <rect x="0.5" y="0.5" width="55" height="55" rx="27.5" fill="white" />
-            <rect x="0.5" y="0.5" width="55" height="55" rx="27.5" stroke="#BABABA" />
-            <path
-                d="M38.56 28.25C38.56 27.47 38.49 26.72 38.36 26H28V30.255H33.92C33.665 31.63 32.89 32.795 31.725 33.575V36.335H35.28C37.36 34.42 38.56 31.6 38.56 28.25Z"
-                fill="#4285F4"
-            />
-            <path
-                d="M27.9999 38.9998C30.9699 38.9998 33.4599 38.0148 35.2799 36.3348L31.7249 33.5748C30.7399 34.2348 29.4799 34.6248 27.9999 34.6248C25.1349 34.6248 22.7099 32.6898 21.8449 30.0898H18.1699V32.9398C19.9799 36.5348 23.6999 38.9998 27.9999 38.9998Z"
-                fill="#34A853"
-            />
-            <path
-                d="M21.845 30.0896C21.625 29.4296 21.5 28.7246 21.5 27.9996C21.5 27.2746 21.625 26.5696 21.845 25.9096V23.0596H18.17C17.4 24.5924 16.9993 26.2842 17 27.9996C17 29.7746 17.425 31.4546 18.17 32.9396L21.845 30.0896Z"
-                fill="#FBBC05"
-            />
-            <path
-                d="M27.9999 21.375C29.6149 21.375 31.0649 21.93 32.2049 23.02L35.3599 19.865C33.4549 18.09 30.9649 17 27.9999 17C23.6999 17 19.9799 19.465 18.1699 23.06L21.8449 25.91C22.7099 23.31 25.1349 21.375 27.9999 21.375Z"
-                fill="#EA4335"
-            />
-        </svg>
-    );
-};
-
-export default GoogleLogo;
