--- src/app/(auth)/log-in/page.tsx
@@ -1,11 +1,0 @@
-import LogInContainer from './_components/LogInContainer';
-
-const LogInPage = () => {
-    return (
-        <main className="flex flex-col items-center justify-center h-screen overflow-y-scroll">
-            <LogInContainer />
-        </main>
-    );
-};
-
-export default LogInPage;
